LayoutTests/http/tests/security/mixedContent/insecure-iframe-in-main-frame-UpgradeMixedContent.html

<!doctype html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

window.data = null;
window.addEventListener("message", function (e) {
    window.data = e.data;
    shouldBe("window.data", "'blocked'");

    if (window.testRunner)
        testRunner.notifyDone();
}, false);

</script>
<p>This test opens a window that loads an insecure iframe.  We should trigger a
mixed content callback and block the request because the main frame in the
window is HTTPS but is displaying insecure content.</p>
<script>
onload = function() {
    window.open("https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-frame.html");
}
</script>
</body>
</html>
