LayoutTests/http/tests/security/mixedContent/insecure-iframe-in-main-frame-UpgradeMixedContent-expected.txt

CONSOLE MESSAGE: [blocked] The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-frame.html requested insecure content from http://127.0.0.1:8080/security/mixedContent/resources/boring.html. This content was blocked and must be served over HTTPS.

PASS window.data is 'blocked'
This test opens a window that loads an insecure iframe. We should trigger a mixed content callback and block the request because the main frame in the window is HTTPS but is displaying insecure content.
